Added a couple explicit integer conversions

+++ src/lrmfit.f
@@ -234,7 +234,7 @@ C
             ENDDO
             cpiy=lprob(bx + off)
             IF(calcc) THEN
-               ipp=500d0*cpiy + 1d0
+               ipp=INT(500d0*cpiy + 1d0)
                if(.NOT. normwt) incobs=w + .5D0
                ftable(ipp,iy + 1)=ftable(ipp,iy + 1) + incobs
             ENDIF
@@ -274,7 +274,7 @@ C     Add to first and second derivatives
          ENDIF
 C     COMPUTE EXCEEDENCE PROBABILITIES AND CHECK FOR DIVERGENCE
          IF(calcc) THEN
-            ipp=500d0*prob(bx + beta(mid) + off) + 1d0
+            ipp=INT(500d0*prob(bx + beta(mid) + off) + 1d0)
             if(.NOT. normwt)incobs=w + .5D0
             ftable(ipp,iy + 1)=ftable(ipp,iy + 1) + incobs
          ENDIF
